requirements.txt
transformers
torch
pillow
requests
accelerate
